% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_code}
\alias{add_code}
\title{Add code block}
\usage{
add_code(
  content,
  code,
  language = "r",
  caption = NULL,
  filename = NULL,
  tabgroup = NULL
)
}
\arguments{
\item{content}{A content_collection or viz_collection object}

\item{code}{Code content}

\item{language}{Programming language for syntax highlighting}

\item{caption}{Optional caption}

\item{filename}{Optional filename to display}

\item{tabgroup}{Optional tabgroup for organizing content (character vector for nested tabs)}
}
\value{
Updated content_collection
}
\description{
Add code block
}
