% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_creation.R
\name{add_content}
\alias{add_content}
\title{Add content collection to a page}
\usage{
add_content(page, content)
}
\arguments{
\item{page}{A page_object created by create_page()}

\item{content}{A content collection (from create_viz/create_content)}
}
\value{
The updated page_object
}
\description{
Add a pre-built content collection (from create_viz/create_content) to a page.
Use this when you have complex content built separately.
}
