% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_creation.R
\name{add_pages}
\alias{add_pages}
\title{Add multiple pages to a dashboard}
\usage{
add_pages(proj, ...)
}
\arguments{
\item{proj}{A dashboard_project object}

\item{...}{One or more page_objects to add}
}
\value{
The updated dashboard_project object
}
\description{
Adds one or more page objects to a dashboard.
}
\examples{
\dontrun{
home <- create_page("Home", is_landing_page = TRUE) \%>\%
  add_text("# Welcome!")

analysis <- create_page("Analysis", data = gss, type = "bar") \%>\%
  add_viz(x_var = "degree", title = "Education") \%>\%
  add_viz(x_var = "race", title = "Race")

create_dashboard(title = "My Dashboard") \%>\%
  add_pages(home, analysis) \%>\%
  generate_dashboard()
}
}
