% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_card.R
\name{add_sparkline_card_row}
\alias{add_sparkline_card_row}
\title{Start a sparkline card row}
\usage{
add_sparkline_card_row(content, tabgroup = NULL, show_when = NULL)
}
\arguments{
\item{content}{Content collection or page_object}

\item{tabgroup}{Optional tabgroup}

\item{show_when}{Optional conditional display formula}
}
\description{
Creates a container for sparkline cards displayed side by side.
}
\keyword{internal}
