% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_ui.R
\name{card_ui}
\alias{card_ui}
\title{Function to encapsulate a shiny tag into a card}
\usage{
card_ui(title, body, df_graph, r = NULL)
}
\arguments{
\item{title}{title of the card}

\item{body}{shiny tag to encapsulate}

\item{df_graph}{A graph dashboarder dataframe}

\item{r}{r internal list (advanced use)}
}
\value{
A shiny tag encapsulated into a card
}
\description{
Function to encapsulate a shiny tag into a card
}
