% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_ui.R
\name{col_ui}
\alias{col_ui}
\title{shiny tag encapsulation into a column}
\usage{
col_ui(section, sect_val, r)
}
\arguments{
\item{section}{shiny tag to encapsulate}

\item{sect_val}{a list of section parameters}

\item{r}{r internal list (advanced use)}
}
\value{
shiny tag encapsulated into a column
}
\description{
shiny tag encapsulation into a column
}
