% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_collection.R
\name{combine_viz}
\alias{combine_viz}
\title{Combine visualization collections}
\usage{
combine_viz(...)
}
\arguments{
\item{...}{One or more viz_collection objects to combine}
}
\value{
A combined viz_collection
}
\description{
This function has been superseded by \code{\link[=combine_content]{combine_content()}}. It still works
but we recommend using \code{combine_content()} for new code as it handles
all content types and attributes more reliably.
}
\examples{
\dontrun{
viz1 <- create_viz() \%>\% add_viz(type = "histogram", x_var = "age")
viz2 <- create_viz() \%>\% add_viz(type = "histogram", x_var = "income")
combined <- combine_viz(viz1, viz2)  # Combines both
}
}
