% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gauge_steps.R
\name{compute_gauge_steps}
\alias{compute_gauge_steps}
\title{Compute step parameter for gauge}
\usage{
compute_gauge_steps(indic_val, theme_var)
}
\arguments{
\item{indic_val}{df associated indicator values}

\item{theme_var}{the list of bslib variables}
}
\value{
a step parameter for a gauge
}
\description{
Compute step parameter for gauge
}
