% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_indicator.R
\name{compute_indicator}
\alias{compute_indicator}
\title{create an indicator plotly graph}
\usage{
compute_indicator(fig, indic_val, theme_var)
}
\arguments{
\item{fig}{a default plotly indicator}

\item{indic_val}{df associated indicator values}

\item{theme_var}{the list of bslib variables}
}
\value{
a new plotly indicator
}
\description{
create an indicator plotly graph
}
