% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph.R
\name{create_graph}
\alias{create_graph}
\title{create a default plotly graph}
\usage{
create_graph(theme_var)
}
\arguments{
\item{theme_var}{the list of bslib variables}
}
\value{
a new default plotly graph
}
\description{
create a default plotly graph
}
