% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indicator.R
\name{create_indicator}
\alias{create_indicator}
\title{create a default indicator plotly graph}
\usage{
create_indicator(theme_var)
}
\arguments{
\item{theme_var}{the list of bslib variables}
}
\value{
a new default indicator plotly graph
}
\description{
create a default indicator plotly graph
}
