% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_card.R
\name{dash_card}
\alias{dash_card}
\title{gridcard with tooltips and processing of default values}
\usage{
dash_card(area, val, r, pattern)
}
\arguments{
\item{area}{gridcard area}

\item{val}{ui default value}

\item{r}{r internal list (advanced uses)}

\item{pattern}{default pattern to identify r code}
}
\value{
a gridcard with tooltips and processing of default values
}
\description{
gridcard with tooltips and processing of default values
}
