% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_el_val.R
\name{def_el_val}
\alias{def_el_val}
\title{Define list of element parameters}
\usage{
def_el_val(df_graph, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{df_graph}{graphical element dasboardr dataframe}

\item{r}{internal r list}

\item{default_pattern}{default pattern to evaluate}
}
\value{
a list of element parameters
}
\description{
Define list of element parameters
}
