% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_gauge_val_server.R
\name{def_gauge_val_server}
\alias{def_gauge_val_server}
\title{Define list of infobox parameters}
\usage{
def_gauge_val_server(df, r = NULL, pattern = "^\%r\%")
}
\arguments{
\item{df}{graphical element dasboardr dataframe}

\item{r}{internal r list}

\item{pattern}{default pattern to evaluate}
}
\value{
a list of infobox parameters
}
\description{
Define list of infobox parameters
}
