% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_gauge_val_ui.R
\name{def_gauge_val_ui}
\alias{def_gauge_val_ui}
\title{Define list of ui gauge parameters}
\usage{
def_gauge_val_ui(df, r = NULL, pattern = "^\%r\%")
}
\arguments{
\item{df}{dasboardr dataframe of an element}

\item{r}{internal r list}

\item{pattern}{default pattern to evaluate}
}
\value{
a list of gauge parameters
}
\description{
Define list of ui gauge parameters
}
