% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_indic_val.R
\name{def_indic_val}
\alias{def_indic_val}
\title{Define list of indicator parameters}
\usage{
def_indic_val(df_graph, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{df_graph}{graphical element dasboardr dataframe}

\item{r}{internal r list}

\item{default_pattern}{default pattern to evaluate}
}
\value{
a list of indicator parameters
}
\description{
Define list of indicator parameters
}
