% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_infobox_val.R
\name{def_infobox_val}
\alias{def_infobox_val}
\title{Define list of infobox parameters}
\usage{
def_infobox_val(df_graph, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{df_graph}{graphical element dasboardr dataframe}

\item{r}{internal r list}

\item{default_pattern}{default pattern to evaluate}
}
\value{
a list of infobox parameters
}
\description{
Define list of infobox parameters
}
