% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_infobox_val_ui.R
\name{def_infobox_val_ui}
\alias{def_infobox_val_ui}
\title{Define list of ui infobox parameters}
\usage{
def_infobox_val_ui(df, r = NULL, pattern = "^\%r\%")
}
\arguments{
\item{df}{dasboardr dataframe of an element}

\item{r}{internal r list}

\item{pattern}{default pattern to evaluate}
}
\value{
a list of infobox parameters
}
\description{
Define list of ui infobox parameters
}
