% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_nav_val.R
\name{def_nav_val}
\alias{def_nav_val}
\title{Define list of nav parameters}
\usage{
def_nav_val(df_graph, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{df_graph}{graphical element dasboardr dataframe}

\item{r}{internal r list}

\item{default_pattern}{default pattern to evaluate}
}
\value{
a list of nav parameters
}
\description{
Define list of nav parameters
}
