% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_row_val.R
\name{def_row_val}
\alias{def_row_val}
\title{Define list of row parameters}
\usage{
def_row_val(df_row, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{df_row}{graphical element dasboardr dataframe}

\item{r}{internal r list}

\item{default_pattern}{default pattern to evaluate}
}
\value{
a list of row parameters
}
\description{
Define list of row parameters
}
