% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_sect_val.R
\name{def_sect_val}
\alias{def_sect_val}
\title{Define list of card parameters}
\usage{
def_sect_val(df_sect, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{df_sect}{section dasboardr dataframe}

\item{r}{internal r list}

\item{default_pattern}{default pattern to evaluate}
}
\value{
a list of card parameters
}
\description{
Define list of card parameters
}
