% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_ui_val.R
\name{def_ui_val}
\alias{def_ui_val}
\title{Define list of ui infobox parameters}
\usage{
def_ui_val(df, r = NULL, pattern = "^\%r\%")
}
\arguments{
\item{df}{dasboardr dataframe of an element}

\item{r}{internal r list}

\item{pattern}{default pattern to evaluate}
}
\value{
a list of infobox parameters
}
\description{
Define list of ui infobox parameters
}
