% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated-viz}
\alias{deprecated-viz}
\alias{create_histogram}
\alias{create_bar}
\alias{create_stackedbar}
\alias{create_stackedbars}
\alias{create_timeline}
\alias{create_heatmap}
\alias{create_scatter}
\alias{create_map}
\alias{create_treemap}
\title{Deprecated Visualization Functions}
\usage{
create_histogram(...)

create_bar(...)

create_stackedbar(...)

create_stackedbars(...)

create_timeline(...)

create_heatmap(...)

create_scatter(...)

create_map(...)

create_treemap(...)
}
\description{
These functions have been renamed to use the \verb{viz_*} prefix for clarity.
Please use the new names:
\itemize{
\item \code{create_histogram()} -> \code{viz_histogram()}
\item \code{create_bar()} -> \code{viz_bar()}
\item \code{create_stackedbar()} -> \code{viz_stackedbar()}
\item \code{create_stackedbars()} -> \code{viz_stackedbars()}
\item \code{create_timeline()} -> \code{viz_timeline()}
\item \code{create_heatmap()} -> \code{viz_heatmap()}
\item \code{create_scatter()} -> \code{viz_scatter()}
\item \code{create_map()} -> \code{viz_map()}
\item \code{create_treemap()} -> \code{viz_treemap()}
}

The old names will continue to work but are deprecated and will be
removed in a future version.
}
\keyword{internal}
