% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.apply_legend_ggplot}
\alias{.apply_legend_ggplot}
\title{Apply legend position to a ggiraph/ggplot chart}
\usage{
.apply_legend_ggplot(p, legend_position, default_show = TRUE)
}
\arguments{
\item{p}{ggplot object (before girafe())}

\item{legend_position}{Character: "top", "bottom", "left", "right", or "none"}

\item{default_show}{Logical: whether to show legend by default}
}
\description{
Apply legend position to a ggiraph/ggplot chart
}
\keyword{internal}
