% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.apply_legend_plotly}
\alias{.apply_legend_plotly}
\title{Apply legend position to a plotly chart}
\usage{
.apply_legend_plotly(p, legend_position, default_show = TRUE)
}
\arguments{
\item{p}{plotly chart object}

\item{legend_position}{Character: "top", "bottom", "left", "right", or "none"}

\item{default_show}{Logical: whether to show legend by default}
}
\description{
Apply legend position to a plotly chart
}
\keyword{internal}
