% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.apply_tooltip_to_hc}
\alias{.apply_tooltip_to_hc}
\title{Apply processed tooltip configuration to a highchart}
\usage{
.apply_tooltip_to_hc(hc, tooltip_result)
}
\arguments{
\item{hc}{A highchart object}

\item{tooltip_result}{Result from .process_tooltip_config()}
}
\value{
Modified highchart object
}
\description{
Apply processed tooltip configuration to a highchart
}
\keyword{internal}
