% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.assert_backend_supported}
\alias{.assert_backend_supported}
\title{Assert backend support for a viz type}
\usage{
.assert_backend_supported(viz_type, backend)
}
\arguments{
\item{viz_type}{Character. Viz type name (e.g., "bar")}

\item{backend}{Character. Backend name}
}
\description{
Assert backend support for a viz type
}
\keyword{internal}
