% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\docType{data}
\name{.backend_capabilities}
\alias{.backend_capabilities}
\title{Registry of supported backends per visualization type}
\format{
An object of class \code{list} of length 20.
}
\usage{
.backend_capabilities
}
\description{
Used to enforce hard errors when a backend does not support a viz type.
}
\keyword{internal}
