% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_creation.R
\name{.blocks_contain_show_when}
\alias{.blocks_contain_show_when}
\title{Recursively check if any content block has show_when set}
\usage{
.blocks_contain_show_when(blocks)
}
\arguments{
\item{blocks}{List of content blocks (may contain nested content collections)}
}
\value{
Logical
}
\description{
Recursively check if any content block has show_when set
}
\keyword{internal}
