% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_bar_formatter}
\alias{.build_bar_formatter}
\title{Build bar chart formatter}
\usage{
.build_bar_formatter(pre, suf, xsuf, context)
}
\description{
Build bar chart formatter
}
\keyword{internal}
