% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_boxplot_formatter}
\alias{.build_boxplot_formatter}
\title{Build boxplot formatter}
\usage{
.build_boxplot_formatter(pre, suf, context)
}
\description{
Build boxplot formatter
}
\keyword{internal}
