% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_default_formatter}
\alias{.build_default_formatter}
\title{Build default formatter}
\usage{
.build_default_formatter(pre, suf, context)
}
\description{
Build default formatter
}
\keyword{internal}
