% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_density_formatter}
\alias{.build_density_formatter}
\title{Build density formatter}
\usage{
.build_density_formatter(pre, suf, context)
}
\description{
Build density formatter
}
\keyword{internal}
