% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_heatmap_formatter}
\alias{.build_heatmap_formatter}
\title{Build heatmap formatter}
\usage{
.build_heatmap_formatter(pre, suf, context)
}
\description{
Build heatmap formatter
}
\keyword{internal}
