% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_histogram_formatter}
\alias{.build_histogram_formatter}
\title{Build histogram formatter}
\usage{
.build_histogram_formatter(pre, suf, xsuf, context)
}
\description{
Build histogram formatter
}
\keyword{internal}
