% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{.build_margin_style}
\alias{.build_margin_style}
\title{Build CSS margin style string from individual margin parameters}
\usage{
.build_margin_style(mt = NULL, mr = NULL, mb = NULL, ml = NULL)
}
\description{
Build CSS margin style string from individual margin parameters
}
\keyword{internal}
