% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_scatter_formatter}
\alias{.build_scatter_formatter}
\title{Build scatter formatter}
\usage{
.build_scatter_formatter(pre, suf, context)
}
\description{
Build scatter formatter
}
\keyword{internal}
