% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_stackedbar_formatter}
\alias{.build_stackedbar_formatter}
\title{Build stacked bar formatter}
\usage{
.build_stackedbar_formatter(pre, suf, xsuf, context)
}
\description{
Build stacked bar formatter
}
\keyword{internal}
