% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_timeline_formatter}
\alias{.build_timeline_formatter}
\title{Build timeline formatter}
\usage{
.build_timeline_formatter(pre, suf, context)
}
\description{
Build timeline formatter
}
\keyword{internal}
