% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{.build_treemap_formatter}
\alias{.build_treemap_formatter}
\title{Build treemap formatter}
\usage{
.build_treemap_formatter(pre, suf, context)
}
\description{
Build treemap formatter
}
\keyword{internal}
