% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard_new.R
\name{.check_for_icons}
\alias{.check_for_icons}
\title{Check if any icons are used in the dashboard}
\usage{
.check_for_icons(proj)
}
\arguments{
\item{proj}{A dashboard_project object}
}
\value{
Logical indicating if icons are present
}
\description{
Internal function to detect if iconify shortcodes are present
in the dashboard content.
}
\keyword{internal}
