% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination_controls.R
\name{.create_pagination_nav_html}
\alias{.create_pagination_nav_html}
\title{Internal function to create pagination navigation HTML}
\usage{
.create_pagination_nav_html(page_num, total_pages, base_name, position)
}
\description{
Internal function to create pagination navigation HTML
}
\keyword{internal}
