% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.detect_widget_backend}
\alias{.detect_widget_backend}
\title{Detect widget backend from class}
\usage{
.detect_widget_backend(widget)
}
\description{
Detect widget backend from class
}
\keyword{internal}
