% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_card_block}
\alias{.generate_card_block}
\title{Generate card block markdown}
\usage{
.generate_card_block(block)
}
\arguments{
\item{block}{Card content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for card content blocks
}
\keyword{internal}
