% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.generate_chart_id}
\alias{.generate_chart_id}
\title{Generate a unique chart id for dashboardr widgets}
\usage{
.generate_chart_id(prefix = "dr_chart")
}
\description{
Generate a unique chart id for dashboardr widgets
}
\keyword{internal}
