% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{.generate_checkbox_html}
\alias{.generate_checkbox_html}
\title{Generate HTML for checkbox input}
\usage{
.generate_checkbox_html(
  input_id,
  label,
  filter_var,
  options,
  default_selected,
  width,
  align,
  inline,
  size = "md",
  help = NULL,
  disabled = FALSE,
  columns = NULL,
  stacked = FALSE,
  stacked_align = "center",
  group_align = "left",
  ncol = NULL,
  nrow = NULL
)
}
\description{
Generate HTML for checkbox input
}
\keyword{internal}
