% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_divider_block}
\alias{.generate_divider_block}
\title{Generate divider block markdown}
\usage{
.generate_divider_block(block)
}
\arguments{
\item{block}{Divider content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for divider content blocks
}
\keyword{internal}
