% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_DT_block}
\alias{.generate_DT_block}
\title{Generate DT datatable block markdown}
\usage{
.generate_DT_block(block)
}
\arguments{
\item{block}{DT content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for DT::datatable content blocks
}
\keyword{internal}
