% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_gt_block}
\alias{.generate_gt_block}
\title{Generate gt table block markdown}
\usage{
.generate_gt_block(block)
}
\arguments{
\item{block}{GT table content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for gt table content blocks
}
\keyword{internal}
