% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_hc_block}
\alias{.generate_hc_block}
\title{Generate highcharter block markdown}
\usage{
.generate_hc_block(block)
}
\arguments{
\item{block}{Highcharter content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for custom highcharter content blocks
}
\keyword{internal}
