% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_image_block}
\alias{.generate_image_block}
\title{Generate image block markdown}
\usage{
.generate_image_block(block)
}
\arguments{
\item{block}{Image content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for image content blocks
}
\keyword{internal}
