% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_metric_block}
\alias{.generate_metric_block}
\title{Generate metric block markdown}
\usage{
.generate_metric_block(block)
}
\arguments{
\item{block}{Metric content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for metric/value box content blocks
}
\keyword{internal}
