% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{.generate_number_html}
\alias{.generate_number_html}
\title{Generate HTML for number input}
\usage{
.generate_number_html(
  input_id,
  label,
  filter_var,
  min,
  max,
  step,
  value,
  width,
  align,
  size = "md",
  help = NULL,
  disabled = FALSE
)
}
\description{
Generate HTML for number input
}
\keyword{internal}
