% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_quote_block}
\alias{.generate_quote_block}
\title{Generate quote block markdown}
\usage{
.generate_quote_block(block)
}
\arguments{
\item{block}{Quote content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for blockquote content blocks
}
\keyword{internal}
