% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_reactable_block}
\alias{.generate_reactable_block}
\title{Generate reactable block markdown}
\usage{
.generate_reactable_block(block)
}
\arguments{
\item{block}{Reactable content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for reactable content blocks
}
\keyword{internal}
