% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{.generate_select_html}
\alias{.generate_select_html}
\title{Generate HTML for select input}
\usage{
.generate_select_html(
  input_id,
  label,
  type,
  filter_var,
  options,
  default_selected,
  placeholder,
  width,
  align,
  size = "md",
  help = NULL,
  disabled = FALSE
)
}
\description{
Generate HTML for select input
}
\keyword{internal}
