% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{.generate_slider_html}
\alias{.generate_slider_html}
\title{Generate HTML for slider input}
\usage{
.generate_slider_html(
  input_id,
  label,
  filter_var,
  min,
  max,
  step,
  value,
  show_value,
  width,
  align,
  labels = NULL,
  size = "md",
  help = NULL,
  disabled = FALSE
)
}
\description{
Generate HTML for slider input
}
\keyword{internal}
