% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_sparkline_card_row_block}
\alias{.generate_sparkline_card_row_block}
\title{Generate sparkline card row block markdown}
\usage{
.generate_sparkline_card_row_block(block, backend = NULL)
}
\arguments{
\item{block}{Sparkline card row content block}
}
\value{
Character vector of markdown lines
}
\description{
Generate sparkline card row block markdown
}
\keyword{internal}
