% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{.generate_switch_html}
\alias{.generate_switch_html}
\title{Generate HTML for switch input}
\usage{
.generate_switch_html(
  input_id,
  label,
  filter_var,
  value,
  width,
  align,
  toggle_series = NULL,
  override = FALSE,
  size = "md",
  help = NULL,
  disabled = FALSE
)
}
\description{
Generate HTML for switch input
}
\keyword{internal}
