% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_table_block}
\alias{.generate_table_block}
\title{Generate table block markdown}
\usage{
.generate_table_block(block)
}
\arguments{
\item{block}{Table content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for table content blocks (data frames using knitr::kable)
}
\keyword{internal}
