% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{.generate_text_html}
\alias{.generate_text_html}
\title{Generate HTML for text input}
\usage{
.generate_text_html(
  input_id,
  label,
  filter_var,
  value,
  placeholder,
  width,
  align,
  size = "md",
  help = NULL,
  disabled = FALSE
)
}
\description{
Generate HTML for text input
}
\keyword{internal}
