% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_generation.R
\name{.generate_widget_block}
\alias{.generate_widget_block}
\title{Generate widget block markdown}
\usage{
.generate_widget_block(block)
}
\arguments{
\item{block}{Widget content block}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown for generic htmlwidget content blocks
(plotly, leaflet, echarts4r, ggiraph, etc.)
}
\keyword{internal}
