% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.page_has_ggiraph}
\alias{.page_has_ggiraph}
\title{Detect ggiraph usage in a page (viz specs or widgets)}
\usage{
.page_has_ggiraph(page, proj_backend = NULL)
}
\description{
Detect ggiraph usage in a page (viz specs or widgets)
}
\keyword{internal}
