% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard_new.R
\name{.progress_bar}
\alias{.progress_bar}
\title{Show progress bar}
\usage{
.progress_bar(current, total, label = "", show_progress = TRUE)
}
\arguments{
\item{current}{Current step}

\item{total}{Total steps}

\item{label}{Optional label}

\item{show_progress}{Whether to show progress}
}
\description{
Show progress bar
}
\keyword{internal}
