% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard_new.R
\name{.progress_header}
\alias{.progress_header}
\title{Show progress header}
\usage{
.progress_header(title, show_progress = TRUE)
}
\arguments{
\item{title}{Header title}

\item{show_progress}{Whether to show progress}
}
\description{
Show progress header
}
\keyword{internal}
