% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard_new.R
\name{.progress_msg}
\alias{.progress_msg}
\title{Show custom progress message}
\usage{
.progress_msg(msg, icon = "▪", show_progress = TRUE)
}
\arguments{
\item{msg}{Message to display}

\item{icon}{Emoji or symbol to prefix}

\item{show_progress}{Whether to show progress}
}
\description{
Show custom progress message
}
\keyword{internal}
