% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_display.R
\name{.progress_step}
\alias{.progress_step}
\title{Show custom progress step}
\usage{
.progress_step(
  msg,
  elapsed = NULL,
  show_progress = TRUE,
  is_last = FALSE,
  use_page_style = FALSE
)
}
\arguments{
\item{msg}{Step message}

\item{elapsed}{Optional elapsed time in seconds}

\item{show_progress}{Whether to show progress}
}
\description{
Show custom progress step
}
\keyword{internal}
