% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.register_dt_widget}
\alias{.register_dt_widget}
\title{Register a DT widget for filtering}
\usage{
.register_dt_widget(widget, table_id, filter_vars = NULL, data = NULL)
}
\description{
Register a DT widget for filtering
}
\keyword{internal}
