% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.register_reactable_widget}
\alias{.register_reactable_widget}
\title{Register a reactable widget for filtering}
\usage{
.register_reactable_widget(widget, table_id, filter_vars = NULL, data = NULL)
}
\description{
Register a reactable widget for filtering
}
\keyword{internal}
