% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_card.R
\name{.sparkline_aggregate}
\alias{.sparkline_aggregate}
\title{Aggregate data for sparkline}
\usage{
.sparkline_aggregate(data, x_var, y_var, agg, filter_expr)
}
\description{
Aggregate data for sparkline
}
\keyword{internal}
