% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_creation.R
\name{.specs_contain_show_when}
\alias{.specs_contain_show_when}
\title{Recursively check if any spec in a list has show_when set}
\usage{
.specs_contain_show_when(specs)
}
\arguments{
\item{specs}{List of visualization specs (may contain nested_children)}
}
\value{
Logical
}
\description{
Recursively check if any spec in a list has show_when set
}
\keyword{internal}
