% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\docType{data}
\name{.viz_param_registry}
\alias{.viz_param_registry}
\title{Registry of visualization parameter requirements}
\format{
An object of class \code{list} of length 11.
}
\usage{
.viz_param_registry
}
\description{
Internal list defining required and optional parameters for each viz type,
along with which parameters represent column names that should be validated.
}
\keyword{internal}
