% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{end_input_row}
\alias{end_input_row}
\title{End an input row}
\usage{
end_input_row(row_container)
}
\arguments{
\item{row_container}{Input row container object}
}
\value{
The parent content_collection for further piping
}
\description{
Closes an input row and returns to the parent content collection.
Must be called after add_input_row() and all add_input() calls.
}
\examples{
\dontrun{
content <- create_content() \%>\%
  add_input_row() \%>\%
    add_input(input_id = "filter1", filter_var = "var1", options = c("A", "B")) \%>\%
    add_input(input_id = "filter2", filter_var = "var2", options = c("X", "Y")) \%>\%
  end_input_row() \%>\%
  add_text("Content after the input row...")
}
}
