% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{end_layout_column}
\alias{end_layout_column}
\title{End a manual layout column}
\usage{
end_layout_column(column_container)
}
\arguments{
\item{column_container}{A layout_column_container created by \code{add_layout_column()}.}
}
\value{
The parent content_collection or page_object.
}
\description{
End a manual layout column
}
