% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{end_layout_row}
\alias{end_layout_row}
\title{End a manual layout row}
\usage{
end_layout_row(row_container)
}
\arguments{
\item{row_container}{A layout_row_container created by \code{add_layout_row()}.}
}
\value{
The parent layout_column_container.
}
\description{
End a manual layout row
}
