% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{end_sidebar}
\alias{end_sidebar}
\title{End a sidebar}
\usage{
end_sidebar(sidebar_container)
}
\arguments{
\item{sidebar_container}{Sidebar container object created by add_sidebar()}
}
\value{
The parent content_collection or page_object for further piping
}
\description{
Closes a sidebar container and returns to the parent content collection.
Must be called after add_sidebar() and all content additions.
}
\examples{
\dontrun{
content <- create_content() \%>\%
  add_sidebar() \%>\%
    add_text("## Filters") \%>\%
    add_input(input_id = "filter1", filter_var = "var1", options = c("A", "B")) \%>\%
  end_sidebar() \%>\%
  add_text("Content after the sidebar...")
}
}
