% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_card.R
\name{end_sparkline_card_row}
\alias{end_sparkline_card_row}
\title{End a sparkline card row}
\usage{
end_sparkline_card_row(row_container)
}
\arguments{
\item{row_container}{Sparkline card row container}
}
\description{
Closes a sparkline card row and returns to the parent content.
}
\keyword{internal}
