% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bslib_theme.R
\name{extract_bslib_theme}
\alias{extract_bslib_theme}
\title{Extract values of selected bslib variables}
\usage{
extract_bslib_theme(session)
}
\arguments{
\item{session}{shiny session}
}
\value{
List of values for selected bslib variables
}
\description{
Extract values of selected bslib variables
}
