% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_name.R
\name{extract_var_name}
\alias{extract_var_name}
\title{Extract variable name}
\usage{
extract_var_name()
}
\value{
a list of variable names organized by thematic
}
\description{
Extract variable name
}
