% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_compute.R
\name{gauge_compute}
\alias{gauge_compute}
\title{create an indicator plotly graph}
\usage{
gauge_compute(fig, indic_val, theme_var)
}
\arguments{
\item{fig}{a default plotly indicator}

\item{indic_val}{df associated indicator values}

\item{theme_var}{the list of bslib variables}
}
\value{
a new plotly indicator
}
\description{
create an indicator plotly graph
}
