% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_create.R
\name{gauge_create}
\alias{gauge_create}
\title{create a default indicator plotly graph}
\usage{
gauge_create(theme_var)
}
\arguments{
\item{theme_var}{the list of bslib variables}
}
\value{
a new default indicator plotly graph
}
\description{
create a default indicator plotly graph
}
