% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_steps.R
\name{gauge_steps}
\alias{gauge_steps}
\title{Compute step parameter for gauge}
\usage{
gauge_steps(indic_val, theme_var)
}
\arguments{
\item{indic_val}{df associated indicator values}

\item{theme_var}{the list of bslib variables}
}
\value{
a step parameter for a gauge
}
\description{
Compute step parameter for gauge
}
