% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_default_theme.R
\name{graph_default_theme}
\alias{graph_default_theme}
\title{apply a default plotly theme for graph}
\usage{
graph_default_theme(graph, theme_var)
}
\arguments{
\item{graph}{a graph to which apply the default theme}

\item{theme_var}{the list of bslib variables}
}
\value{
a graph with the default plotly theme applied
}
\description{
apply a default plotly theme for graph
}
