% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_card2.R
\name{grid_card2}
\alias{grid_card2}
\title{Augmented grid-positioned card element}
\usage{
grid_card2(..., title_class = NULL, footer = NULL, footer_class = NULL)
}
\arguments{
\item{...}{Parameters of grid_card function of the gridlayout package}

\item{title_class}{class for card-title}

\item{footer}{tag to include in the card footer}

\item{footer_class}{class for card-footer}
}
\value{
Augmented grid-positioned card element
}
\description{
Augmented grid-positioned card element
}
