% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\name{html_iframe}
\alias{html_iframe}
\title{Create an iframe embed}
\usage{
html_iframe(url, height = "500px", width = "100\%", style = NULL)
}
\arguments{
\item{url}{URL to embed.}

\item{height}{CSS height value. Default "500px".}

\item{width}{CSS width value. Default "100\%".}

\item{style}{Optional additional CSS styles.}
}
\value{
An htmltools tag object.
}
\description{
Create an iframe embed
}
