% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Function to import data from excel spreadsheet}
\usage{
import_data(path)
}
\arguments{
\item{path}{path to the excel spreadsheet}
}
\value{
a dataframe of data from excel spreadsheet
}
\description{
Function to import data from excel spreadsheet
}
