% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_ext_fct.R
\name{int_ext_fct}
\alias{int_ext_fct}
\title{Identify and compute internal/external graph characteristics}
\usage{
int_ext_fct(x, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{x}{value}

\item{r}{intern list of variables}

\item{default_pattern}{default pattern to identify r code in dashboardr dataframe}
}
\value{
real value (internal or external)
}
\description{
Identify and compute internal/external graph characteristics
}
