% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_helpers.R
\name{is_content_block}
\alias{is_content_block}
\title{Check if object is a content block}
\usage{
is_content_block(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a content block
}
\description{
Internal helper to check if an object is a content block.
Content blocks represent individual pieces of content like text,
images, callouts, etc.
}
\keyword{internal}
