% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{is_tooltip}
\alias{is_tooltip}
\title{Check if object is a tooltip configuration}
\usage{
is_tooltip(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is a tooltip configuration
}
\keyword{internal}
