% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{merge_collections}
\alias{merge_collections}
\title{Merge two content/viz collections}
\usage{
merge_collections(c1, c2)
}
\arguments{
\item{c1}{First collection}

\item{c2}{Second collection}
}
\value{
Merged content_collection
}
\description{
Internal function to merge two collections into one.
}
