% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboard_ui.R
\name{mod_dashboard_ui}
\alias{mod_dashboard_ui}
\title{dashboard ui part}
\usage{
mod_dashboard_ui(id, df, r = NULL, pattern = "^\%r\%")
}
\arguments{
\item{id}{module id}

\item{df}{A dashboardr dataframe}

\item{r}{r internal list (advanced use)}

\item{pattern}{default pattern to identify r code in dashboardr dataframe}
}
\value{
dashboard ui part
}
\description{
dashboard ui part
}
