% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboardr_server.R
\name{mod_dashboardr_server}
\alias{mod_dashboardr_server}
\title{dashboardr Server part}
\usage{
mod_dashboardr_server(id, df, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{id}{A module id}

\item{df}{A dashboarder dataframe}

\item{r}{r internal list (advanced use)}

\item{default_pattern}{default pattern to identify r code in dashboardr dataframe}
}
\value{
a dashboard
}
\description{
dashboardr Server part
}
