% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboardr_ui.R
\name{mod_dashboardr_ui}
\alias{mod_dashboardr_ui}
\title{dashboardr ui part}
\usage{
mod_dashboardr_ui(id, df, r = NULL)
}
\arguments{
\item{id}{module id}

\item{df}{A dashboarder dataframe}

\item{r}{r internal list (advanced use)}
}
\value{
a dashboard
}
\description{
dashboardr ui part
}
