% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_el_server.R
\name{mod_el_server}
\alias{mod_el_server}
\title{dashboardr Server part}
\usage{
mod_el_server(id, df_graph, r = NULL, theme_var, default_pattern = "^\%r\%")
}
\arguments{
\item{id}{A module id}

\item{df_graph}{A graph dashboarder dataframe}

\item{r}{r internal list (advanced use)}

\item{theme_var}{Shiny app theme list of interesting variables}

\item{default_pattern}{default pattern to identify r code in dashboardr dataframe}
}
\value{
a dashboard
}
\description{
dashboardr Server part
}
