% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_el_ui.R
\name{mod_el_ui}
\alias{mod_el_ui}
\title{Create a ui shiny tag for plotly graph}
\usage{
mod_el_ui(id, df_graph, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{id}{plot id}

\item{df_graph}{A dashboarder graph dataframe}

\item{r}{r internal list (advanced use)}

\item{default_pattern}{default_pattern = default_pattern}
}
\value{
a ui shiny tag for plotly graph
}
\description{
Create a ui shiny tag for plotly graph
}
