% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gauge_server.R
\name{mod_gauge_server}
\alias{mod_gauge_server}
\title{gauge server part}
\usage{
mod_gauge_server(id, df, r = NULL, theme_var, pattern = "^\%r\%")
}
\arguments{
\item{id}{module id}

\item{df}{A graph dashboarder dataframe}

\item{r}{r internal list (advanced use)}

\item{theme_var}{bslib theme variables}

\item{pattern}{default pattern to identify r code in dashboardr dataframe}
}
\value{
an gauge server
}
\description{
gauge server part
}
