% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gauge_ui.R
\name{mod_gauge_ui}
\alias{mod_gauge_ui}
\title{gauge ui part}
\usage{
mod_gauge_ui(id, df, r = NULL, pattern = "^\%r\%")
}
\arguments{
\item{id}{module id}

\item{df}{A graph dashboarder dataframe}

\item{r}{r internal list (advanced use)}

\item{pattern}{default pattern to identify r code in dashboardr dataframe}
}
\value{
a gauge ui
}
\description{
gauge ui part
}
