% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_ui.R
\name{mod_graph_ui}
\alias{mod_graph_ui}
\title{Create a ui shiny tag for plotly graph}
\usage{
mod_graph_ui(id, df_graph, r = NULL)
}
\arguments{
\item{id}{plot id}

\item{df_graph}{A dashboarder graph dataframe}

\item{r}{r internal list (advanced use)}
}
\value{
a ui shiny tag for plotly graph
}
\description{
Create a ui shiny tag for plotly graph
}
