% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infobox_ui.R
\name{mod_infobox_ui}
\alias{mod_infobox_ui}
\title{infobox ui part}
\usage{
mod_infobox_ui(id, df_graph, r = NULL, default_pattern = "^\%r\%")
}
\arguments{
\item{id}{module id}

\item{df_graph}{A graph dashboarder dataframe}

\item{r}{r internal list (advanced use)}

\item{default_pattern}{default pattern to identify r code in dashboardr dataframe}
}
\value{
an infobox ui
}
\description{
infobox ui part
}
