% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_ui.R
\name{nav_ui}
\alias{nav_ui}
\title{shiny tag encapsulation into a nav}
\usage{
nav_ui(graph, df_sect, df_graph, r, default_pattern)
}
\arguments{
\item{graph}{a graph to encapsulate}

\item{df_sect}{A section dashboardr dataframe}

\item{df_graph}{A graph dashboardr dataframe}

\item{r}{r internal list (advanced use)}

\item{default_pattern}{default pattern to evaluate}
}
\value{
shiny tag encapsulated into a nav
}
\description{
shiny tag encapsulation into a nav
}
