% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{plot_to_html}
\alias{plot_to_html}
\title{Convert a Plot Object to HTML}
\usage{
plot_to_html(plot)
}
\arguments{
\item{plot}{A plot object that can be rendered as HTML (e.g., \code{highchart} object).}
}
\value{
A character string containing the HTML representation of the plot,
or "NULL plot" if the input was NULL.
}
\description{
This function takes a plot object (presumably a \code{highcharter} or
\code{htmlwidget} object) and wraps it in HTML \code{div} tags. It is
intended for embedding plots in HTML documents, such as those
generated by Quarto or R Markdown.
}
\details{
If the input \code{plot} is NULL, it prints a message to the console and
returns the string "NULL plot". Otherwise, it converts the plot to an
HTML tag list using \code{htmltools::tagList} and then wraps this in a
\code{div} with class "flex-item2". The result is returned as a character string.
}
\examples{
# Assuming 'my_hc_plot' is a highchart object:
# html_code <- plot_to_html(my_hc_plot)
# cat(html_code)

# Handling a NULL plot:
# plot_to_html(NULL)

Dependencies: htmltools, shiny (tags$div)
}
