% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{+.viz_collection}
\alias{+.viz_collection}
\title{Combine viz collections using + operator}
\usage{
\method{+}{viz_collection}(e1, e2)
}
\arguments{
\item{e1}{First viz_collection}

\item{e2}{Second viz_collection}
}
\value{
A merged collection containing items from both
}
\description{
Combine viz collections using + operator
}
