% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_helpers.R
\name{print.dashboardr_tooltip}
\alias{print.dashboardr_tooltip}
\title{Print method for tooltip configurations}
\usage{
\method{print}{dashboardr_tooltip}(x, ...)
}
\arguments{
\item{x}{A dashboardr_tooltip object}

\item{...}{Ignored}
}
\description{
Print method for tooltip configurations
}
