% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_collection.R
\name{print.dashboardr_widget}
\alias{print.dashboardr_widget}
\title{Print method for dashboardr_widget - opens in viewer}
\usage{
\method{print}{dashboardr_widget}(x, ...)
}
\description{
Print method for dashboardr_widget - opens in viewer
}
