% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_creation.R
\name{print.page_object}
\alias{print.page_object}
\title{Print method for page objects}
\usage{
\method{print}{page_object}(x, ...)
}
\arguments{
\item{x}{A page_object to print}

\item{...}{Additional arguments (currently ignored)}
}
\description{
Print method for page objects
}
