% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{render_input_row}
\alias{render_input_row}
\title{Render a row of input widgets}
\usage{
render_input_row(inputs, style = "boxed", align = "center")
}
\arguments{
\item{inputs}{List of input specifications (each should have the same
parameters as render_input)}

\item{style}{Visual style: "boxed" (default) or "inline" (compact)}

\item{align}{Alignment: "center" (default), "left", or "right"}
}
\value{
HTML output
}
\description{
Creates HTML for a horizontal row of input widgets.
}
