% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_card.R
\name{render_sparkline_card_row}
\alias{render_sparkline_card_row}
\title{Render a row of sparkline cards}
\usage{
render_sparkline_card_row(data, cards, backend = "echarts4r")
}
\arguments{
\item{data}{Data frame (page data)}

\item{cards}{List of card specifications}

\item{backend}{Chart backend}
}
\description{
Render a row of sparkline cards
}
\keyword{internal}
