% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box_render.R
\name{render_value_box_row}
\alias{render_value_box_row}
\title{Render a row of value boxes}
\usage{
render_value_box_row(boxes)
}
\arguments{
\item{boxes}{List of value box specifications, each containing title, value, bg_color, logo_url, logo_text}
}
\description{
Render a row of value boxes
}
