% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_ui.R
\name{row_ui}
\alias{row_ui}
\title{shiny tag encapsulation into a row}
\usage{
row_ui(row, df_row, r, default_pattern)
}
\arguments{
\item{row}{a row to encapsulate}

\item{df_row}{A row dashboardr dataframe}

\item{r}{r internal list (advanced use)}

\item{default_pattern}{default pattern to evaluate}
}
\value{
shiny tag encapsulated into a row
}
\description{
shiny tag encapsulation into a row
}
