% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_collection.R
\name{save_widget}
\alias{save_widget}
\title{Save widget as self-contained HTML}
\usage{
save_widget(widget, file, selfcontained = TRUE)
}
\arguments{
\item{widget}{A dashboardr widget created with preview(output = "widget")}

\item{file}{Path to save the HTML file}

\item{selfcontained}{Whether to embed all dependencies (default: TRUE)}
}
\value{
Invisibly returns the file path
}
\description{
Save widget as self-contained HTML
}
\examples{
\dontrun{
widget <- my_viz \%>\% preview(output = "widget")
save_widget(widget, "my_chart.html")
}
}
