% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_collection.R
\name{show_structure}
\alias{show_structure}
\title{Show collection structure (even with data attached)}
\usage{
show_structure(x)
}
\arguments{
\item{x}{A content_collection or viz_collection object}
}
\value{
In knitr: formatted HTML output. In console: invisible(x) after printing.
}
\description{
Forces display of the collection structure instead of rendering visualizations.
Useful when you want to inspect the structure of a collection that has data
attached, or when documenting the collection's organization.
}
\examples{
\dontrun{
# In a vignette or R Markdown, pipe into print() to see the tree
# even when data is attached:
create_viz(data = mtcars, type = "bar") \%>\%
  add_viz(x_var = "cyl", title = "Cylinders") \%>\%
  print()
}
}
