% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_df.R
\name{test_df}
\alias{test_df}
\title{Test the conformity of necessary variables in dashboardr dataframe}
\usage{
test_df(df, r = NULL, check_var)
}
\arguments{
\item{df}{Dashboardr dataframe}

\item{r}{r internal value}

\item{check_var}{Vector of variables to test}
}
\value{
several error messages (if any)
}
\description{
Test the conformity of necessary variables in dashboardr dataframe
}
