% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{vecc}
\alias{vecc}
\title{Create a Concatenated String}
\usage{
vecc(name, num)
}
\arguments{
\item{name}{A character string.}

\item{num}{A number (will be coerced to character by \code{paste0}).}
}
\value{
A character string which is the concatenation of \code{name} and \code{num}.
}
\description{
This is a simple utility function that concatenates a name (string)
and a number.
}
\examples{
vecc("Variable", 1)   # Returns "Variable1"
vecc("Q", 10)         # Returns "Q10"
}
