% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_callout}
\alias{add_callout}
\title{Add callout box}
\usage{
add_callout(
  content,
  text,
  type = c("note", "tip", "warning", "caution", "important"),
  title = NULL,
  icon = NULL,
  collapse = FALSE
)
}
\arguments{
\item{content}{A content_collection or viz_collection object}

\item{text}{Callout content}

\item{type}{Callout type (note/tip/warning/caution/important)}

\item{title}{Optional title}

\item{icon}{Optional icon}

\item{collapse}{Whether callout is collapsible}
}
\value{
Updated content_collection
}
\description{
Add callout box
}
