% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_card}
\alias{add_card}
\title{Add card}
\usage{
add_card(content, text, title = NULL, footer = NULL)
}
\arguments{
\item{content}{A content_collection or viz_collection object}

\item{text}{Card content}

\item{title}{Card title}

\item{footer}{Optional card footer}
}
\value{
Updated content_collection
}
\description{
Add card
}
