% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_divider}
\alias{add_divider}
\title{Add horizontal divider}
\usage{
add_divider(content, style = "default")
}
\arguments{
\item{content}{A content_collection object}

\item{style}{Divider style ("default", "thick", "dashed", "dotted")}
}
\value{
Updated content_collection
}
\description{
Add horizontal divider
}
