% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_DT}
\alias{add_DT}
\title{Add DT datatable}
\usage{
add_DT(content, table_data, options = NULL, ...)
}
\arguments{
\item{content}{A content_collection object}

\item{table_data}{A DT datatable object (from DT::datatable()) OR a data frame/matrix (will be auto-converted)}

\item{options}{List of DT options (only used if passing a data frame)}

\item{...}{Additional arguments passed to DT::datatable() (only used if passing a data frame)}
}
\value{
Updated content_collection
}
\description{
Add DT datatable
}
\examples{
\dontrun{
# Option 1: Pass a styled DT object
my_dt <- DT::datatable(
  mtcars,
  options = list(pageLength = 10),
  filter = 'top',
  rownames = FALSE
)

content <- create_content() \%>\%
  add_DT(my_dt)
  
# Option 2: Pass a data frame with options
content <- create_content() \%>\%
  add_DT(mtcars, options = list(pageLength = 5, scrollX = TRUE))
}
}
