% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_html}
\alias{add_html}
\title{Add raw HTML content}
\usage{
add_html(content, html, tabgroup = NULL, show_when = NULL)
}
\arguments{
\item{content}{Content collection object}

\item{html}{Raw HTML string}

\item{tabgroup}{Optional tabgroup for organizing content (character vector for nested tabs)}

\item{show_when}{One-sided formula controlling conditional display based on input values.}
}
\description{
Add raw HTML content
}
