% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_html}
\alias{add_html}
\title{Add raw HTML content}
\usage{
add_html(content, html)
}
\arguments{
\item{content}{Content collection object}

\item{html}{Raw HTML string}
}
\description{
Add raw HTML content
}
