% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_quote}
\alias{add_quote}
\title{Add a blockquote}
\usage{
add_quote(content, quote, attribution = NULL, cite = NULL, tabgroup = NULL)
}
\arguments{
\item{content}{Content collection object}

\item{quote}{Quote text}

\item{attribution}{Optional attribution/source}

\item{cite}{Optional citation URL}

\item{tabgroup}{Optional tabgroup for organizing content (character vector for nested tabs)}
}
\description{
Add a blockquote
}
