% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_reactable}
\alias{add_reactable}
\title{Add reactable table}
\usage{
add_reactable(content, reactable_object)
}
\arguments{
\item{content}{A content_collection object}

\item{reactable_object}{A reactable object (from reactable::reactable()) OR a data frame (will be auto-converted)}
}
\value{
Updated content_collection
}
\description{
Add reactable table
}
\examples{
\dontrun{
# Option 1: Pass a styled reactable object
my_table <- reactable::reactable(
  mtcars,
  columns = list(mpg = reactable::colDef(name = "MPG")),
  searchable = TRUE,
  striped = TRUE
)

content <- create_content() \%>\%
  add_reactable(my_table)
  
# Option 2: Pass a data frame (auto-converted with defaults)
content <- create_content() \%>\%
  add_reactable(mtcars)
}
}
