% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_spacer}
\alias{add_spacer}
\title{Add vertical spacer}
\usage{
add_spacer(content, height = "2rem")
}
\arguments{
\item{content}{A content_collection or viz_collection object}

\item{height}{Height (CSS unit, e.g. "2rem", "50px")}
}
\value{
Updated content_collection
}
\description{
Add vertical spacer
}
