% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_table}
\alias{add_table}
\title{Add generic table (data frame)}
\usage{
add_table(content, table_object, caption = NULL)
}
\arguments{
\item{content}{A content_collection object}

\item{table_object}{A data frame or tibble}

\item{caption}{Optional caption}
}
\value{
Updated content_collection
}
\description{
Add generic table (data frame)
}
