% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tooltip.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Add tooltip to section title}
\usage{
add_tooltip(df_sect)
}
\arguments{
\item{df_sect}{df_sect object}
}
\value{
a html icon with tooltip
}
\description{
Add tooltip to section title
}
