% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_collection.R
\name{add_video}
\alias{add_video}
\title{Add video}
\usage{
add_video(content, src, caption = NULL, width = NULL, height = NULL)
}
\arguments{
\item{content}{A content_collection object}

\item{src}{Video source URL or path}

\item{caption}{Optional caption}

\item{width}{Optional width}

\item{height}{Optional height}
}
\value{
Updated content_collection
}
\description{
Add video
}
