% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_ui.R
\name{card_ui}
\alias{card_ui}
\title{Function to encapsulate a shiny tag into a card}
\usage{
card_ui(body, sect_val, r = NULL)
}
\arguments{
\item{body}{shiny tag to encapsulate}

\item{sect_val}{a list of section parameters}

\item{r}{r internal list (advanced use)}
}
\value{
A shiny tag encapsulated into a card
}
\description{
Function to encapsulate a shiny tag into a card
}
