% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_collection.R
\name{combine_viz}
\alias{combine_viz}
\title{Combine visualization collections}
\usage{
combine_viz(...)
}
\arguments{
\item{...}{One or more viz_collection objects to combine}
}
\value{
A combined viz_collection
}
\description{
Alternative function to combine viz_collection objects.
}
\examples{
\dontrun{
viz1 <- create_viz() \%>\% add_viz(type = "histogram", x_var = "age")
viz2 <- create_viz() \%>\% add_viz(type = "histogram", x_var = "income")
combined <- combine_viz(viz1, viz2)  # Combines both
}
}
