% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{.generate_button_group_html}
\alias{.generate_button_group_html}
\title{Generate HTML for button group input}
\usage{
.generate_button_group_html(
  input_id,
  label,
  filter_var,
  options,
  default_selected,
  width,
  align,
  size = "md",
  help = NULL,
  disabled = FALSE
)
}
\description{
Generate HTML for button group input
}
\keyword{internal}
