% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_display.R
\name{.progress_section}
\alias{.progress_section}
\title{Show progress section}
\usage{
.progress_section(title, show_progress = TRUE)
}
\arguments{
\item{title}{Section title}

\item{show_progress}{Whether to show progress}
}
\description{
Show progress section
}
\keyword{internal}
