% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboardr_server.R
\name{mod_dashboardr_server}
\alias{mod_dashboardr_server}
\title{dashboardr Server part}
\usage{
mod_dashboardr_server(id, df)
}
\arguments{
\item{id}{A module id}

\item{df}{A dashboarder dataframe}
}
\value{
a dashboard
}
\description{
dashboardr Server part
}
