% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_server.R
\name{mod_graph_server}
\alias{mod_graph_server}
\title{dashboardr Server part}
\usage{
mod_graph_server(id, df_graph, r = NULL)
}
\arguments{
\item{id}{A module id}

\item{df_graph}{A graph dashboarder dataframe}

\item{r}{r internal list (advanced use)}
}
\value{
a dashboard
}
\description{
dashboardr Server part
}
