% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_df.R
\name{norm_df}
\alias{norm_df}
\title{Function to normalize a dashboardr dataframe}
\usage{
norm_df(df, r)
}
\arguments{
\item{df}{A dashboardr dataframe}

\item{r}{r internal list (advanced use)}
}
\value{
A normalized dashboardr dataframe
}
\description{
Function to normalize a dashboardr dataframe
}
