% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_df.R
\name{test_df}
\alias{test_df}
\title{Function to test the validity of a dashboardr dataframe}
\usage{
test_df(df, r)
}
\arguments{
\item{df}{A dashboardr dataframe}

\item{r}{r internal list (advanced use)}
}
\value{
An error if the dashboardr dataframe is not valid
}
\description{
Function to test the validity of a dashboardr dataframe
}
